chore: add validation address

+++ src/domain/usecases/address/deleteAddress.js
@@ -8,6 +8,14 @@ class DeleteAddress {
       throw new Error('O idAddress é obrigatório')
     }
 
+    const address = await this.addressRepository.getAddressById({ idAddress })
+    if (!address) {
+      throw new Error('Endereço não encontrado')
+    }
+    if (address.isMain) {
+      throw new Error('Não é possível excluir o endereço principal')
+    }
+
     const response = await this.addressRepository.deleteAddressById({
       idAddress
     })
